deleted PrintOnConsole.cs and moved its methods to SearchDrinkConsoleUi.cs
